+++ common/include/common/HttpRequest.h
@@ -97,6 +97,10 @@ public:
     }
 
     void send(std::string_view body = "");
+
+    // returns the HTTP status code
+    int send_no_check(std::string_view body = "");
+    
     size_t read(void* buf, size_t buf_size);
     std::optional<std::string> get_header(std::string_view name);
 };