+++ game_patch/fflink/fflink.cpp
@@ -1,0 +1,18 @@
+#include "fflink.h"
+#include "fflink_session.h"
+#include "fflink_utils.h"
+
+namespace fflink {
+
+void do_patch()
+{
+    session_do_patch();
+}
+
+void do_frame()
+{
+    drain_pending_main_thread_tasks();
+    drain_pending_console();
+}
+
+} // namespace fflink