+++ game_patch/fflink/fflink.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace fflink {
+
+void do_patch();
+void do_frame();
+
+} // namespace fflink